+++ lib/src/tweet_video.dart
@@ -11,6 +11,7 @@ class TweetVideo extends StatefulWidget {
     this.autoPlay = false,
     this.enableFullscreen = true,
     this.videoHighQuality = true,
+    this.controlBarColor,
   }) : super(key: key);
 
   final TweetVM tweetVM;
@@ -18,6 +19,7 @@ class TweetVideo extends StatefulWidget {
   final bool autoPlay;
   final bool enableFullscreen;
   final bool? videoHighQuality;
+  final Color? controlBarColor;
 
   @override
   _TweetVideoState createState() => _TweetVideoState();
@@ -44,6 +46,7 @@ class _TweetVideoState extends State<TweetVideo>
       },
       aspectRatio: widget.tweetVM.getDisplayTweet().videoAspectRatio!,
       controlsConfiguration: BetterPlayerControlsConfiguration(
+        controlBarColor: widget.controlBarColor ?? Colors.black87,
         enablePlaybackSpeed: false,
         enableSkips: false,
         enableMute: !widget.tweetVM.getDisplayTweet().hasGif,