+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Fixed url launcher deprecation - use canLaunchUrl and launchUrl instead of canLaunch and launch
+- Added videoControlBarBgColor for edit the video player control bar background
+
 
 ## [4.0.2] - 28.05.2022
 ### Changed